Move DownloadOqOutputTask to a separate file in tasks folder
